Fix accessor's min/max

+++ Core/Scripts/Format/glTFBuffer.cs
@@ -87,8 +87,8 @@ namespace UniGLTF
         public string type;
         public glComponentType componentType;
         public int count;
-        public double[] max;
-        public double[] min;
+        public float[] max;
+        public float[] min;
 
         public string ToJson()
         {
